Add new scripts
Former-commit-id: a8ee71f9196d695b2738eaaf77e7ce7b5bb59c67

+++ bin/hold.csh
@@ -1,6 +1,7 @@
-rm hold
+#!/bin/tcsh -f
+if (-r hold) rm hold
 touch hold
-foreach f ( `condq | grep -w H | awk '{print $9}'` )
+foreach f ( `condor_q -wide -submitter fisyak | grep -w H | awk '{print $9}'` )
 ls -1d ${PWD}/${f} >> hold
 end
 set l = `awk -F_ '{print $2}' hold | awk -F\. '{print $1}' | xargs | sed -e 's/ /,/g'`
